refactor(vlm): split vision bridge router
